small code improvements + version bump for publish on crates.io
