+++ apps/cargo-scout-audit/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cargo-scout-audit"
-version = "0.2.5"
+version = "0.2.6"
 edition = "2021"
 authors = [
     "Agustin Aon <agustin.aon@coinfabrik.com>",
