+++ apps/cargo-scout-audit/src/utils/mod.rs
@@ -3,5 +3,6 @@ pub mod cargo_package;
 pub mod command;
 pub mod config;
 pub mod detectors;
+pub mod detectors_info;
 pub mod env;
 pub mod rustup;
