+++ scout-audit-dylint-linting/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "scout-audit-dylint-linting"
-version = "3.0.0"
+version = "3.0.1"
 authors = ["Samuel E. Moelius III <sam@moeli.us>"]
 description = "Utilities for writing Dylint libraries"
 edition = "2021"
